execute {dGEMcovid}	R Documentation

Execute the Study

Description

Execute the Study

Usage

execute(
  databaseDetails,
  siteId = "site_name",
  outputFolder,
  createCohorts = F,
  createData = F,
  sampleSize = NULL,
  createControl = F,
  siteIds = "",
  leadSiteNextStep = F,
  runInitialize = F,
  runDerive = F,
  runEstimate = F,
  runSynthesize = F,
  verbosity = "INFO",
  alphaStart = "20201101",
  alphaEnd = "20210201",
  deltaStart = "20210701",
  deltaEnd = "20211001"
)

Arguments

databaseDetails
The connection details and OMOP CDM details. Created using PatientLevelPrediction::createDatabaseDetails.

siteId
The name of your site (can be the university name, site with a number, etc.) tnis needs to be shared with the study administrator

outputFolder
Name of local folder to place results - make sure control is in here; make sure to use forward slashes (/). Do not use a folder on a network drive since this greatly impacts performance.

createCohorts
Create the cohortTable table with the target population and outcome cohorts?

createData
Create the labelled data set (this is required before runAnalysis)

sampleSize
The number of patients in the target cohort to sample (if NULL uses all patients)

createControl
(for the lead site only) Run this code to create the study control

siteIds
(for the lead site only) vector with the names of all the sites contributing to the study. Required when createControl = TRUE

runInitialize
Runs the initialization step - this require downloading the json control from https://pda-ota.pdamethods.org/

runDerive
Runs the derive step - this require downloading the updated json control from https://pda-ota.pdamethods.org/

runEstimate
Runs the estimate step - this require downloading the updated json control from https://pda-ota.pdamethods.org/

runSynthesize
Once the site estimates are returned, it is now possible to apply each model to the data to calculate predictions. This step requires downloading the updated json control from https://pda-ota.pdamethods.org/

verbosity
Sets the level of the verbosity. If the log level is at or higher in priority than the logger threshold, a message will print. The levels are:

DEBUGHighest verbosity showing all debug statements

TRACEShowing information about start and end of steps

INFOShow informative information (Default)

WARNShow warning messages

ERRORShow error messages

FATALBe silent except for fatal errors

@param alphaStart The start date for the alpha wave in the data in "yyyymmdd" format @param alphaEnd The end date for the alpha wave in the data in "yyyymmdd" format @param deltaStart The start date for the delta wave in the data in "yyyymmdd" format @param deltaEnd The end date for the detla wave in the data in "yyyymmdd" format

Details

This function executes the dGEMcovid Study.

Examples

## Not run: 
connectionDetails <- createConnectionDetails(dbms = "postgresql",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")

execute(databaseDetails,
        siteId = 'site_1',
        outputFolder = "c:/temp/study_results", 
        createCohorts = T,
        createData = T,
        sampleSize = 10000,
        createControl = F,
        siteIds = c('site_1', 'site_2'),
        leadSiteNextStep = F,
        runAnalysis = F,
        runSynthesize = F,
        verbosity = "INFO"
        )

## End(Not run)

[Package dGEMcovid version 0.0.1 Index]